+++ apps/docs/src/app/platform/component-docs/platform-forms/platform-input/platform-input-header/platform-input-header.component.ts
@@ -1,0 +1,7 @@
+import { Component } from '@angular/core';
+
+@Component({
+    selector: 'fd-platform-input-header',
+    templateUrl: './platform-input-header.component.html'
+})
+export class PlatformInputHeaderComponent {}